requirements.txt
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
matplotlib
pandas
python-dotenv
tqdm
tweepy
yfinance
